WIP: Move inferring of spectrum path to parsing psm list?
